Improve GCE Detection Accuracy
It turns out that BIOS vendor being Google exists for things that are
not only GCE VMs, which can cause odd configs as a result of what was
expected to be configured for a physical host to have virtual host
based configs applied. This makes it more accurate in that regard.